+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Improved
 - Improved performance of diagnostics.
 - Parsed quests are stored for accesses from both diagnostics and DocumentSymbolProvider.
+- Updated list of actions and effects.
 
 ## 0.6.0